SupportSprint/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.6
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
gunicorn==20.1.0
